moving old tests from converter to a test file
